remove some warnings (#343)
* remove some warnings

* Wdouble promotion

+++ actuator.h
@@ -73,28 +73,28 @@ class Actuator {
       float output_wrap_adjustment = 0.0;
       if (startup_param_.output_encoder_rollover > 0) {
          if (status.output_position > startup_param_.output_encoder_rollover) {
-            logger.log_printf("Output pos (%2.3f) > (%2.3f), sub 2pi", status.output_position, startup_param_.output_encoder_rollover);
+            logger.log_printf("Output pos (%2.3f) > (%2.3f), sub 2pi", (double)status.output_position, (double)startup_param_.output_encoder_rollover);
             output_wrap_adjustment = -2*M_PI;
          } else if (status.output_position < (startup_param_.output_encoder_rollover-2*M_PI)) {
-            logger.log_printf("Output pos (%2.3f) < (%2.3f), add 2pi", status.output_position, startup_param_.output_encoder_rollover-2*M_PI);
+            logger.log_printf("Output pos (%2.3f) < (%2.3f), add 2pi", (double)status.output_position, (double)(startup_param_.output_encoder_rollover-2*M_PI));
             output_wrap_adjustment = 2*M_PI;
          } else {
-            logger.log_printf("Output pos (%2.3f), rollover (%f), no adjust", status.output_position, startup_param_.output_encoder_rollover);
+            logger.log_printf("Output pos (%2.3f), rollover (%f), no adjust", (double)status.output_position, (double)startup_param_.output_encoder_rollover);
          }
       } else {
          if (status.output_position < startup_param_.output_encoder_rollover) {
-            logger.log_printf("Output pos (%2.3f) < (%2.3f), add 2pi", status.output_position, startup_param_.output_encoder_rollover);
+            logger.log_printf("Output pos (%2.3f) < (%2.3f), add 2pi", (double)status.output_position, (double)startup_param_.output_encoder_rollover);
             output_wrap_adjustment = 2*M_PI;
          } else if (status.output_position > (startup_param_.output_encoder_rollover+2*M_PI)) {
-            logger.log_printf("Output pos (%2.3f) > (%2.3f), sub 2pi", status.output_position, startup_param_.output_encoder_rollover+2*M_PI);
+            logger.log_printf("Output pos (%2.3f) > (%2.3f), sub 2pi", (double)status.output_position, (double)(startup_param_.output_encoder_rollover+2*M_PI));
             output_wrap_adjustment = -2*M_PI;
          } else {
-            logger.log_printf("Output pos (%2.3f), rollover (%f), no adjust", status.output_position, startup_param_.output_encoder_rollover);
+            logger.log_printf("Output pos (%2.3f), rollover (%f), no adjust", (double)status.output_position, (double)startup_param_.output_encoder_rollover);
          }
       }
       main_loop_.adjust_output_encoder(output_wrap_adjustment);
       status.output_position += output_wrap_adjustment;
-      logger.log_printf("Output pos after (%f) adjustment: %f", output_wrap_adjustment, status.output_position);
+      logger.log_printf("Output pos after (%f) adjustment: %f", (double)output_wrap_adjustment, (double)status.output_position);
       switch(startup_param_.motor_encoder_startup) {
          default:
          case StartupParam::ENCODER_ZERO:
@@ -119,11 +119,11 @@ class Actuator {
               - (status.fast_loop.motor_position.position);
             float motor_bias_rounded = roundf(motor_bias_from_output/round_by)*(round_by) + startup_motor_bias_;
             logger.log("Encoder bias from output with correction");
-            logger.log_printf("output_position_to_motor_position: %f", output_position_to_motor_position(status.output_position));
+            logger.log_printf("output_position_to_motor_position: %f", (double)output_position_to_motor_position(status.output_position));
             logger.log_printf("Output position: %f, motor_position: %f, motor mechanical position: %f", 
-               status.output_position, status.fast_loop.motor_position.position, status.fast_loop.motor_mechanical_position);
+               (double)status.output_position, (double)status.fast_loop.motor_position.position, (double)status.fast_loop.motor_mechanical_position);
             logger.log_printf("Encoder bias, round_by: %f, motor_bias_from_output: %f, motor_bias_rounded: %f",
-               round_by, motor_bias_from_output, motor_bias_rounded);
+               (double)round_by, (double)motor_bias_from_output, (double)motor_bias_rounded);
             main_loop_.set_motor_encoder_bias(motor_bias_rounded);
             break;
          }
